% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{params,FDEResult-method}
\alias{params,FDEResult-method}
\title{Accessor for analysis parameters in a FDEResult}
\usage{
\S4method{params}{FDEResult}(x, ...)
}
\arguments{
\item{x}{A \code{FDEResult} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{list} of analysis parameters.
}
\description{
Returns the analysis parameter list from a
  \code{FDEResult} object.
}
\examples{
library(S4Vectors)
de <- DataFrame(logFC = c(1.2, -0.5), P.Value = c(0.001, 0.5),
                adj.P.Val = c(0.01, 0.8))
pb <- matrix(rpois(20, 10), nrow = 2, ncol = 10)
rownames(pb) <- paste0("Gene", 1:2)
colnames(pb) <- paste0("Donor", 1:10)
obj <- FDEResult(
    de,
    pb,
    setNames(sqrt(1:10), paste0("Donor", 1:10)),
    params = list(cell_type = "T_cells", condition = "group")
)
params(obj)

}
