% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,FDEResult-method}
\alias{show,FDEResult-method}
\title{Show method for FDEResult}
\usage{
\S4method{show}{FDEResult}(object)
}
\arguments{
\item{object}{A \code{FDEResult} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Prints a compact summary of a \code{FDEResult} object.
}
